+++ docs/CHANGELOG.md
@@ -20,6 +20,9 @@ This page contains the changelog for violet. Please note that you'll encounter t
 
 ## [unreleased]
 
+
+## [v0.11.1] (2026-09-26)
+
 ### Added
 
 * CLI (generate): Added `--root`, largely for debug and E2E test purposes
@@ -160,7 +163,8 @@ This page contains the changelog for violet. Please note that you'll encounter t
 
 Initial tracked commit. Commits prior to this were considered too unstable to bother.
 
-[unreleased]: https://codeberg.org/LunarWatcher/violet/compare/v0.11.0...master
+[unreleased]: https://codeberg.org/LunarWatcher/violet/compare/v0.11.1...master
+[v0.11.1]: https://codeberg.org/LunarWatcher/violet/compare/v0.11.0...v0.11.1
 [v0.11.0]: https://codeberg.org/LunarWatcher/violet/compare/v0.10.0...v0.11.0
 [v0.10.0]: https://codeberg.org/LunarWatcher/violet/compare/v0.9.0...v0.10.0
 [v0.9.0]: https://codeberg.org/LunarWatcher/violet/compare/v0.8.0...v0.9.0